Copy multicast example from tt-metal repo
Plan: The multicast example conducts a matrix-multiplication. I will
change it to convolution example.

Note: The upstream tt-nn binds
`ttnn/cpp/ttnn/operations/conv2d/conv2d.cpp` to `ttnn.conv2d(..)` python
function. `ttnn/test_conv2d.py` example uses `ttnn.conv2d(..)`, but it
causes an exception. After investigating it, I realized that it has a
divided-by-zero error. Unfortunately, when I built tt-metal in a debug
mode on my local machine (to debug it with GDB), my machine kept
freezing. To understand a valid optimization for convolution operation
on TensTorrent machine, the goal of this example is to change the matrix
multiplication with multicast example to a convolution example. For
that, I will not change the structure of multicast. I will focus on the
convolution operation and will extend it to multi-layer of convolutions,
which is also a good research topic how we can restructure data between
convolution layers to efficiently utilize TensTorrent devices.