Fix multicast_advanced example bugs - now it's working
The fifth example is "multicast_advanced". Its number of the input
tiles is the same as the number of Tensix cores. The data move kernel
iterates `i` for [0, number of cores). Each tensix core has `core_id`
from 0 to `number of cores - 1`. When the data move kernel iteration
`i` is its core_id, it becomes a sender core. Otherwise, it's a receiver
core. So all cores become a sender for its turn, and receiver for other
core's turn.

Since the number of input tiles is the same as the number of Tensix
core, we can assume that each core has its own tile in the input.
When it becomes a sender core, it sends its own tile to all other cores
via multicast. Therefore, each core can get all input tiles i.e., its
own tile and tiles shared by other cores. Each core dump all input tiles
to the output.

The output will have tiles as many as
(`number of cores` x `number of cores`). It will be similar copies of
the input buffer as many as `number of cores`.